+++ hosts/matangi/default.nix
@@ -76,6 +76,12 @@
         integrationLibs = true;
       };
       core.enable = true;
+      fonts = {
+        enable = true;
+        nerdFonts = true;
+        iconFonts = true;
+        systemFonts = true;
+      };
       development = {
         enable = false;
         linters = false;